ci: update rust version (#7883)

+++ src/infra/src/file_list/mysql.rs
@@ -308,10 +308,11 @@ SELECT stream, date, file, deleted, min_ts, max_ts, records, original_size, comp
         time_range: Option<(i64, i64)>,
         flattened: Option<bool>,
     ) -> Result<Vec<FileKey>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
-                return Ok(Vec::new());
-            }
+        if let Some((start, end)) = time_range
+            && start == 0
+            && end == 0
+        {
+            return Ok(Vec::new());
         }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
@@ -367,10 +368,11 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         date_range: Option<(String, String)>,
     ) -> Result<Vec<FileKey>> {
-        if let Some((start, end)) = date_range.as_ref() {
-            if start.is_empty() && end.is_empty() {
-                return Ok(Vec::new());
-            }
+        if let Some((start, end)) = date_range.as_ref()
+            && start.is_empty()
+            && end.is_empty()
+        {
+            return Ok(Vec::new());
         }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
@@ -446,10 +448,11 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         time_range: Option<(i64, i64)>,
     ) -> Result<Vec<super::FileId>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
-                return Ok(Vec::new());
-            }
+        if let Some((start, end)) = time_range
+            && start == 0
+            && end == 0
+        {
+            return Ok(Vec::new());
         }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
@@ -562,10 +565,11 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         time_range: Option<(i64, i64)>,
     ) -> Result<Vec<String>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
-                return Ok(Vec::new());
-            }
+        if let Some((start, end)) = time_range
+            && start == 0
+            && end == 0
+        {
+            return Ok(Vec::new());
         }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
@@ -881,16 +885,16 @@ SELECT stream, MIN(min_ts) AS min_ts, MAX(max_ts) AS max_ts, CAST(COUNT(*) AS SI
             "#,
         );
         if deleted {
-            sql = format!("{} AND deleted IS TRUE", sql);
+            sql = format!("{sql} AND deleted IS TRUE");
         }
         let sql = match pk_value {
-            None => format!("{} GROUP BY stream", sql),
-            Some((0, 0)) => format!("{} GROUP BY stream", sql),
+            None => format!("{sql} GROUP BY stream"),
+            Some((0, 0)) => format!("{sql} GROUP BY stream"),
             Some((min, max)) => {
                 if deleted {
-                    format!("{} AND id <= {} GROUP BY stream", sql, max)
+                    format!("{sql} AND id <= {max} GROUP BY stream")
                 } else {
-                    format!("{} AND id > {} AND id <= {} GROUP BY stream", sql, min, max)
+                    format!("{sql} AND id > {min} AND id <= {max} GROUP BY stream")
                 }
             }
         };
@@ -927,7 +931,7 @@ SELECT stream, MIN(min_ts) AS min_ts, MAX(max_ts) AS max_ts, CAST(COUNT(*) AS SI
                 stream_name.unwrap()
             )
         } else {
-            format!("SELECT * FROM stream_stats WHERE org = '{}';", org_id)
+            format!("SELECT * FROM stream_stats WHERE org = '{org_id}';")
         };
         let pool = CLIENT_RO.clone();
         DB_QUERY_NUMS
@@ -949,8 +953,7 @@ SELECT stream, MIN(min_ts) AS min_ts, MAX(max_ts) AS max_ts, CAST(COUNT(*) AS SI
         stream_name: &str,
     ) -> Result<()> {
         let sql = format!(
-            "DELETE FROM stream_stats WHERE stream = '{}/{}/{}';",
-            org_id, stream_type, stream_name
+            "DELETE FROM stream_stats WHERE stream = '{org_id}/{stream_type}/{stream_name}';"
         );
         let pool = CLIENT.clone();
         DB_QUERY_NUMS
@@ -1210,20 +1213,20 @@ UPDATE stream_stats
         // check job status
         let id = ret.try_get::<i64, &str>("id").unwrap_or_default();
         let status = ret.try_get::<i64, &str>("status").unwrap_or_default();
-        if id > 0 && super::FileListJobStatus::from(status) == super::FileListJobStatus::Done {
-            if let Err(e) =
+        if id > 0
+            && super::FileListJobStatus::from(status) == super::FileListJobStatus::Done
+            && let Err(e) =
                 sqlx::query("UPDATE file_list_jobs SET status = ? WHERE status = ? AND id = ?;")
                     .bind(super::FileListJobStatus::Pending)
                     .bind(super::FileListJobStatus::Done)
                     .bind(id)
                     .execute(&mut *tx)
                     .await
-            {
-                if let Err(e) = tx.rollback().await {
-                    log::error!("[MYSQL] rollback update job status error: {e}");
-                }
-                return Err(e.into());
+        {
+            if let Err(e) = tx.rollback().await {
+                log::error!("[MYSQL] rollback update job status error: {e}");
             }
+            return Err(e.into());
         }
         if let Err(e) = tx.commit().await {
             log::error!("[MYSQL] commit add job error: {e}");
@@ -1241,7 +1244,7 @@ UPDATE stream_stats
             lock_id,
             config::get_config().limit.meta_transaction_lock_timeout
         );
-        let unlock_sql = format!("SELECT RELEASE_LOCK('{}')", lock_id);
+        let unlock_sql = format!("SELECT RELEASE_LOCK('{lock_id}')");
         let mut lock_tx = lock_pool.begin().await?;
         DB_QUERY_NUMS.with_label_values(&["get_lock", "", ""]).inc();
         match sqlx::query_scalar::<_, i64>(&lock_sql)
@@ -1964,15 +1967,15 @@ async fn add_column(table: &str, column: &str, data_type: &str) -> Result<()> {
 
     let alert_sql = format!("ALTER TABLE {table} ADD COLUMN {column} {data_type};");
     let mut tx = pool.begin().await?;
-    if let Err(e) = sqlx::query(&alert_sql).execute(&mut *tx).await {
-        if !e.to_string().contains("Duplicate column name") {
-            // Check for the specific MySQL error code for duplicate column
-            log::error!("[MYSQL] Unexpected error in adding column {column}: {}", e);
-            if let Err(e) = tx.rollback().await {
-                log::error!("[MYSQL] Error in rolling back transaction: {}", e);
-            }
-            return Err(e.into());
+    if let Err(e) = sqlx::query(&alert_sql).execute(&mut *tx).await
+        && !e.to_string().contains("Duplicate column name")
+    {
+        // Check for the specific MySQL error code for duplicate column
+        log::error!("[MYSQL] Unexpected error in adding column {column}: {}", e);
+        if let Err(e) = tx.rollback().await {
+            log::error!("[MYSQL] Error in rolling back transaction: {}", e);
         }
+        return Err(e.into());
     }
     if let Err(e) = tx.commit().await {
         log::info!("[MYSQL] Error in committing transaction: {}", e);
